Split admin views into h.admin.views package
