+++ h/admin/views/__init__.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+
+
+def includeme(config):
+    config.include('.index')
+    config.include('.admins')
+    config.include('.badge')
+    config.include('.features')
+    config.include('.groups')
+    config.include('.nipsa')
+    config.include('.staff')
+    config.include('.users')
